roles & permision

+++ app/Http/Controllers/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Category;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryController extends Controller
 {
@@ -15,6 +16,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        Gate::authorize('all_programs');
 
         $categories = Category::with('companies')->withoutTrashed()->latest('id')->paginate(env('PAGINATION_COUNT'));
         return view('admin.categories.index' , compact('categories'));
@@ -27,6 +29,8 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        Gate::authorize('add_program');
+
         return view('admin.categories.create');
     }
 
@@ -113,6 +117,8 @@ class CategoryController extends Controller
      */
     public function destroy($slug)
     {
+        Gate::authorize('delete_program');
+
         $category = Category::whereSlug($slug)->first();
         $category->destroy($category->id);
         return $slug;
@@ -126,6 +132,8 @@ class CategoryController extends Controller
      */
     public function trash()
     {
+        Gate::authorize('recycle_programs');
+
         $categories = Category::onlyTrashed()->latest('id')->paginate(env('PAGINATION_COUNT'));
         return view('admin.categories.trash', compact('categories'));
     }
@@ -138,6 +146,7 @@ class CategoryController extends Controller
      */
     public function restore($slug)
     {
+        Gate::authorize('restore_program');
         $category = Category::onlyTrashed()->whereSlug($slug)->first();
         $category->restore();
         return $slug;
@@ -152,6 +161,7 @@ class CategoryController extends Controller
      */
     public function forcedelete($slug)
     {
+        Gate::authorize('forceDelete_program');
         $category = Category::onlyTrashed()->whereSlug($slug)->first();
         $category->forcedelete();
         return $slug;
